@@ -1,4 +1,4 @@
-// +build 386 amd64
+//go:build 386 || amd64
 
 package fzf
 
