@@ -1,10 +1,11 @@
-// +build !windows
+//go:build !windows
 
 package fzf
 
 import (
 	"os"
 	"os/signal"
+	"strings"
 	"syscall"
 )
 
@@ -18,4 +19,8 @@ func notifyStop(p *os.Process) {
 
 func notifyOnCont(resizeChan chan<- os.Signal) {
 	signal.Notify(resizeChan, syscall.SIGCONT)
+}
+
+func quoteEntry(entry string) string {
+	return "'" + strings.Replace(entry, "'", "'\\''", -1) + "'"
 }
