@@ -1,9 +1,10 @@
-// +build !windows
+//go:build !windows
 
 package tui
 
 import (
 	"io/ioutil"
+	"os"
 	"syscall"
 )
 
@@ -28,4 +29,19 @@ func ttyname() string {
 		}
 	}
 	return ""
+}
+
+// TtyIn returns terminal device to be used as STDIN, falls back to os.Stdin
+func TtyIn() *os.File {
+	in, err := os.OpenFile(consoleDevice, syscall.O_RDONLY, 0)
+	if err != nil {
+		tty := ttyname()
+		if len(tty) > 0 {
+			if in, err := os.OpenFile(tty, syscall.O_RDONLY, 0); err == nil {
+				return in
+			}
+		}
+		return os.Stdin
+	}
+	return in
 }
