@@ -233,6 +233,7 @@ func Run(opts *Options, version string, revision string) {
 		clearCache = util.Once(true)
 		clearSelection = util.Once(true)
 		chunkList.Clear()
+		itemIndex = 0
 		header = make([]string, 0, opts.HeaderLines)
 		go reader.restart(command)
 	}
@@ -241,9 +242,11 @@ func Run(opts *Options, version string, revision string) {
 	for {
 		delay := true
 		ticks++
-		input := func() []rune {
+		input := func(reloaded bool) []rune {
 			paused, input := terminal.Input()
-			if !paused {
+			if reloaded && paused {
+				query = []rune{}
+			} else if !paused {
 				query = input
 			}
 			return query
@@ -254,7 +257,11 @@ func Run(opts *Options, version string, revision string) {
 			}
 			for evt, value := range *events {
 				switch evt {
-
+				case EvtQuit:
+					if reading {
+						reader.terminate()
+					}
+					os.Exit(value.(int))
 				case EvtReadNew, EvtReadFin:
 					if evt == EvtReadFin && nextCommand != nil {
 						restart(*nextCommand)
@@ -269,7 +276,8 @@ func Run(opts *Options, version string, revision string) {
 						opts.Sync = false
 						terminal.UpdateList(PassMerger(&snapshot, opts.Tac), false)
 					}
-					matcher.Reset(snapshot, input(), false, !reading, sort, clearCache())
+					reset := clearCache()
+					matcher.Reset(snapshot, input(reset), false, !reading, sort, reset)
 
 				case EvtSearchNew:
 					var command *string
@@ -288,7 +296,8 @@ func Run(opts *Options, version string, revision string) {
 						break
 					}
 					snapshot, _ := chunkList.Snapshot()
-					matcher.Reset(snapshot, input(), true, !reading, sort, clearCache())
+					reset := clearCache()
+					matcher.Reset(snapshot, input(reset), true, !reading, sort, reset)
 					delay = false
 
 				case EvtSearchProgress:
